NodePublisherNodeSubscriber*Tests: split into dry (without threading) and regular tests
