+++ src/test/kotlin/graphics/scenery/tests/unit/network/NodePublisherNodeSubscriberTest.kt
@@ -1,15 +1,10 @@
 package graphics.scenery.tests.unit.network
 
 import graphics.scenery.Box
-import graphics.scenery.DefaultNode
 import graphics.scenery.Hub
 import graphics.scenery.Scene
-import graphics.scenery.attribute.spatial.DefaultSpatial
-import graphics.scenery.net.NetworkEvent
 import graphics.scenery.net.NodePublisher
 import graphics.scenery.net.NodeSubscriber
-import graphics.scenery.textures.Texture
-import graphics.scenery.utils.Image
 import graphics.scenery.volumes.TransferFunction
 import graphics.scenery.volumes.Volume
 import net.imglib2.type.numeric.integer.UnsignedByteType
@@ -50,6 +45,7 @@ class NodePublisherNodeSubscriberTest {
 
         sub = NodeSubscriber(hub2, ip = "tcp://127.0.0.1", 6660, context = zContext)
         hub2.add(sub)
+
         Thread.sleep(300)
     }
 
@@ -59,35 +55,6 @@ class NodePublisherNodeSubscriberTest {
         sub.close().join()
         t.join()
         zContext.destroy()
-    }
-
-    /**
-     * Instead of network use debug listen and publish
-     */
-    @Test
-    fun integrationSkippingNetwork() {
-
-        val box = Box()
-        box.name = "box"
-        scene1.addChild(box)
-
-        pub.register(scene1)
-        pub.debugPublish(sub::debugListen)
-        sub.networkUpdate(scene2)
-
-        assert(scene2.find("box") != null)
-    }
-
-    @Test
-    fun integrationSceneName() {
-
-        scene1.name = "lol"
-
-        pub.register(scene1)
-        Thread.sleep(1000)
-        sub.networkUpdate(scene2)
-
-        assertEquals("lol", scene2.name)
     }
 
     @Test
@@ -102,93 +69,6 @@ class NodePublisherNodeSubscriberTest {
         sub.networkUpdate(scene2)
 
         assert(scene2.find("box") != null)
-    }
-
-    @Test
-    fun integrationNodeRemoval() {
-        val node1 = DefaultNode("eins")
-        scene1.addChild(node1)
-        pub.register(scene1)
-        pub.debugPublish { sub.debugListen(serializeAndDeserialize(it) as NetworkEvent) }
-        sub.networkUpdate(scene2)
-        assert(scene2.find("eins") != null)
-
-        scene1.removeChild(node1)
-        pub.debugPublish { sub.debugListen(serializeAndDeserialize(it) as NetworkEvent) }
-        sub.networkUpdate(scene1)
-        assert(scene2.find("eins") == null)
-    }
-
-    @Test
-    fun integrationMoveNodeInGraph() {
-        val node1 = DefaultNode("eins")
-        val node2 = DefaultNode("zwei")
-        scene1.addChild(node1)
-        scene1.addChild(node2)
-        pub.register(scene1)
-        pub.debugPublish { sub.debugListen(serializeAndDeserialize(it) as NetworkEvent) }
-        sub.networkUpdate(scene2)
-
-        scene1.removeChild(node2)
-        node1.addChild(node2)
-        pub.debugPublish { sub.debugListen(serializeAndDeserialize(it) as NetworkEvent) }
-        sub.networkUpdate(scene1)
-
-        val zwei = scene2.find("zwei")
-        assertNotNull(zwei)
-        val eins = zwei.parent
-        assertNotNull(eins)
-    }
-
-    @Test
-    fun integrationMoveAttribute() {
-        //stop agent threads
-        val p = pub.close()
-        sub.close().join()
-        p.join()
-
-        val node1 = Box()
-        node1.name = "eins"
-        val node2 = Box()
-        node2.name = "zwei"
-        scene1.addChild(node1)
-        scene1.addChild(node2)
-
-        pub.register(scene1)
-        pub.debugPublish { sub.debugListen(serializeAndDeserialize(it) as NetworkEvent) }
-        sub.networkUpdate(scene2)
-
-        node2.setMaterial(node1.material())
-        pub.debugPublish { sub.debugListen(serializeAndDeserialize(it) as NetworkEvent) }
-        sub.networkUpdate(scene1)
-
-        val eins = scene2.find("eins")
-        val zwei = scene2.find("zwei")
-
-        assert(eins?.materialOrNull() == zwei?.materialOrNull())
-    }
-
-    @Test
-    fun additionalDataTexture() {
-        val box = Box(Vector3f(1.0f, 1.0f, 1.0f))
-        box.name = "box"
-        box.material {
-            textures["diffuse"] = Texture.fromImage(
-                Image.fromResource(
-                    "../../examples/basic/textures/helix.png",
-                    NodePublisherNodeSubscriberTest::class.java
-                )
-            )
-        }
-        scene1.addChild(box)
-
-        pub.register(scene1)
-        pub.debugPublish { sub.debugListen(serializeAndDeserialize(it) as NetworkEvent) }
-        sub.networkUpdate(scene2)
-
-        val box2 = scene2.find("box") as? Box
-        assert(box2?.material()?.textures?.isNotEmpty() ?: false)
-
     }
 
     @Test
@@ -259,18 +139,15 @@ class NodePublisherNodeSubscriberTest {
     }
 
     @Test
-    fun delegateSerializationAndUpdate() {
+    fun integrationSceneName() {
 
-        val node = DefaultNode()
-        val spatial = DefaultSpatial(node)
-        spatial.position = Vector3f(3f, 0f, 0f)
+        scene1.name = "lol"
 
+        pub.register(scene1)
+        Thread.sleep(1000)
+        sub.networkUpdate(scene2)
 
-        val result = serializeAndDeserialize(spatial) as DefaultSpatial
-
-        assertEquals(3f, result.position.x)
-        //should not fail
-        result.position = Vector3f(3f, 0f, 0f)
+        assertEquals("lol", scene2.name)
     }
 }
